Fix x86_64 gcc 4.4.1 build errors:
/smhasher/MurmurHash1.cpp: In function ‘unsigned int MurmurHash1Aligned(const void*, int, unsigned int)’:
/smhasher/MurmurHash1.cpp:82: error: cast from ‘const unsigned char*’ to ‘int’ loses precision

/smhasher/MurmurHash2.cpp: In function ‘uint32_t MurmurHashAligned2(const void*, int, uint32_t)’:
/smhasher/MurmurHash2.cpp:382: error: cast from ‘const unsigned char*’ to ‘int’ loses precision

+++ trunk/MurmurHash2.cpp
@@ -379,7 +379,7 @@ uint32_t MurmurHashAligned2 ( const void * key, int len, uint32_t seed )
 
 	uint32_t h = seed ^ len;
 
-	int align = (int)data & 3;
+	int align = (uint64_t)data & 3;
 
 	if(align && (len >= 4))
 	{
